requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiofiles==23.2.1
python-magic==0.4.27
websockets==12.0
python-dotenv==1.0.0
python-dateutil==2.8.2
orjson==3.9.10
pydantic==2.5.0
python-cors==1.7.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1